wip preparing dependency refactor update
