+++ src/dependency/core/agrupation/base.py
@@ -1,6 +1,8 @@
 from abc import ABC
 
 class ABCModule(ABC):
+    """Abstract Base Class for Module.
+    """
     def __init__(self, name: str) -> None:
         self.name: str = name
