+++ src/dependency/core/agrupation/entrypoint.py
@@ -1,22 +1,23 @@
-import time
 import logging
+import time
 from dependency.core.agrupation.plugin import Plugin
-from dependency.core.injection.container import Container
-from dependency.core.injection.loader import InjectionLoader
+from dependency.core.resolution.container import Container
+from dependency.core.resolution.resolver import InjectionResolver
 _logger = logging.getLogger("DependencyLoader")
-_init_time: float = time.time()
 
 class Entrypoint:
     """Entrypoint for the application.
     """
+    init_time: float = time.time()
+
     def __init__(self, container: Container, plugins: list[Plugin]) -> None:
-        providers = [
+        injectables = [
             provider
             for plugin in plugins
             for provider in plugin.resolve_providers(container)]
 
-        self.loader = InjectionLoader(
+        self.loader = InjectionResolver(
             container=container,
-            providers=providers)
-        self.loader.resolve_dependencies()
-        _logger.info(f"Application started in {time.time() - _init_time} seconds")
+            injectables=injectables)
+        self.loader.resolve_injectables()
+        _logger.info(f"Application started in {time.time() - self.init_time} seconds")
