+++ src/dependency/core/declaration/__init__.py
@@ -1,12 +1,11 @@
+from dependency_injector import providers
 from dependency.core.declaration.component import Component, component
-from dependency.core.declaration.instance import instance, providers
-from dependency.core.declaration.product import Product, product
+from dependency.core.declaration.instance import Instance, instance
 
 __all__ = [
     "Component",
     "component",
+    "Instance",
     "instance",
     "providers",
-    "Product",
-    "product"
 ]
