+++ src/dependency/core/injection/injectable.py
@@ -4,6 +4,8 @@ from dependency_injector import containers, providers
 T = TypeVar('T')
 
 class Injectable:
+    """Injectable Class representing a injectable dependency.
+    """
     def __init__(self,
         component_cls: type,
         provided_cls: type,
@@ -17,7 +19,6 @@ class Injectable:
         self.modules_cls: set[type] = {component_cls}
         self.imports: list["Injectable"] = imports
         self.bootstrap: Optional[Callable[[], Any]] = bootstrap
-
         self.is_resolved: bool = False
 
     @property
@@ -32,20 +33,21 @@ class Injectable:
         self.modules_cls.add(wire_cls)
 
     def wire_provider(self, container: containers.DynamicContainer) -> "Injectable":
+        """Wire the provider with the given container."""
         container.wire(modules=self.modules_cls)
         self.is_resolved = True
         return self
 
     def provide(self) -> Any:
+        """Return an instance from the provider."""
         return self.provider_cls(self.provided_cls)
 
     def prewiring(self) -> None:
+        """Prepare the imports for wiring."""
         for implementation in self.imports:
             implementation.add_wire_cls(self.provided_cls)
 
     def do_bootstrap(self) -> None:
+        """Execute the bootstrap function if it exists."""
         if self.bootstrap is not None:
             self.bootstrap()
-
-    def __hash__(self) -> int:
-        return hash(self.component_cls)
