+++ src/dependency/core/injection/provider.py
@@ -1,9 +1,11 @@
 from typing import Any, Generator, Optional
-from dependency.core2.injection.base import BaseInjection, ContainerInjection
-from dependency.core2.injection.injectable import Injectable
-from dependency.core2.exceptions import DeclarationError
+from dependency.core.injection.base import BaseInjection, ContainerInjection
+from dependency.core.injection.injectable import Injectable
+from dependency.core.exceptions import DeclarationError
 
 class ProviderInjection(BaseInjection):
+    """Provider Injection Class
+    """
     def __init__(self,
         name: str,
         parent: Optional["ContainerInjection"] = None
@@ -14,14 +16,16 @@ class ProviderInjection(BaseInjection):
 
     @property
     def injectable(self) -> Injectable:
+        """Return the injectable instance."""
         if not self.__injectable:
-            raise DeclarationError("Injectable is not set.")
+            raise DeclarationError(f"Implementation for provider {self.name} was not set")
         return self.__injectable
 
     def set_instance(self,
         injectable: Injectable,
         imports: list['ProviderInjection'] = [],
     ) -> None:
+        """Set the injectable instance and its imports."""
         self.__injectable = injectable
         self.__imports = imports
 
@@ -30,6 +34,7 @@ class ProviderInjection(BaseInjection):
         return self.injectable.provide()
 
     def resolve_providers(self) -> Generator[Injectable, None, None]:
+        """Inject all imports into the current injectable."""
         self.injectable.imports = [
             provider.injectable
             for provider in self.__imports
