+++ src/dependency/core/resolution/resolver.py
@@ -1,20 +1,24 @@
 import logging
 from pprint import pformat
-from dependency.core2.injection.injectable import Injectable
-from dependency.core2.resolution.container import Container
-from dependency.core2.resolution.utils import raise_providers_error
+from dependency.core.injection.injectable import Injectable
+from dependency.core.resolution.container import Container
+from dependency.core.resolution.utils import raise_providers_error
 _logger = logging.getLogger("DependencyLoader")
 
 class InjectionResolver:
+    """Injection Resolver Class
+    """
     def __init__(self, container: Container, injectables: list[Injectable]) -> None:
         self.container: Container = container
         self.injectables: list[Injectable] = injectables
 
     def resolve_dependencies(self) -> None:
+        """Resolve all dependencies and initialize them."""
         _logger.info("Dependencies resolved and initialized")
 
     def resolve_injectables(self,
     ) -> list[list[Injectable]]:
+        """Resolve all injectables in layers."""
         unresolved: list[Injectable] = self.injectables
         resolved_layers: list[list[Injectable]] = []
 
@@ -44,6 +48,7 @@ class InjectionResolver:
     def start_implementations(self,
         resolved_layers: list[list[Injectable]],
     ) -> None:
+        """Start all implementations by executing their bootstrap functions."""
         for layer in resolved_layers:
             for implementation in layer:
                 implementation.do_bootstrap()
