--- src/dependency/core2/__init__.py
@@ -1,35 +1,0 @@
-from dependency.core2.agrupation import (
-    Entrypoint,
-    Module,
-    module,
-    Plugin,
-    PluginMeta,
-)
-from dependency.core2.declaration import (
-    Component,
-    component,
-    Instance,
-    instance,
-    providers,
-)
-from dependency.core2.resolution import (
-    Container,
-)
-from dependency.core.exceptions import (
-    DependencyError,
-)
-
-__all__ = [
-    "Entrypoint",
-    "Module",
-    "module",
-    "Plugin",
-    "PluginMeta",
-    "Component",
-    "component",
-    "Instance",
-    "instance",
-    "providers",
-    "Container",
-    "DependencyError",
-]
