--- src/dependency/core2/agrupation/__init__.py
@@ -1,11 +1,0 @@
-from dependency.core2.agrupation.entrypoint import Entrypoint
-from dependency.core2.agrupation.module import Module, module
-from dependency.core2.agrupation.plugin import Plugin, PluginMeta
-
-__all__ = [
-    "Entrypoint",
-    "Module",
-    "module",
-    "Plugin",
-    "PluginMeta",
-]
