--- src/dependency/core2/agrupation/base.py
@@ -1,8 +1,0 @@
-from abc import ABC
-
-class ABCModule(ABC):
-    def __init__(self, name: str) -> None:
-        self.name: str = name
-
-    def __repr__(self) -> str:
-        return self.name
