--- src/dependency/core2/declaration/__init__.py
@@ -1,11 +1,0 @@
-from dependency_injector import providers
-from dependency.core2.declaration.component import Component, component
-from dependency.core2.declaration.instance import Instance, instance
-
-__all__ = [
-    "Component",
-    "component",
-    "Instance",
-    "instance",
-    "providers",
-]
