--- src/dependency/core2/declaration/base.py
@@ -1,27 +1,0 @@
-from abc import ABC, abstractmethod
-from typing import Any
-
-class ABCComponent(ABC):
-    """Abstract base class for all components.
-    """
-    def __init__(self, interface_cls: type) -> None:
-        self.interface_cls: type = interface_cls
-
-    def __repr__(self) -> str:
-        return self.interface_cls.__name__
-
-class ABCInstance(ABC):
-    """Abstract base class for all instances.
-    """
-    def __init__(self, provided_cls: type) -> None:
-        self.provided_cls: type = provided_cls
-
-    def __repr__(self) -> str:
-        return self.provided_cls.__name__
-
-class ABCInjectable(ABC):
-    """Abstract base class for all injectables.
-    """
-    @abstractmethod
-    def provide(self) -> Any:
-        pass
