--- src/dependency/core2/exceptions.py
@@ -1,5 +1,0 @@
-
-class DependencyError(Exception): ...
-
-class DeclarationError(DependencyError): ...
-class ResolutionError(DependencyError): ...
