--- src/dependency/core2/injection/__init__.py
@@ -1,9 +1,0 @@
-from dependency.core2.injection.base import ContainerInjection
-from dependency.core2.injection.provider import ProviderInjection
-from dependency.core2.injection.injectable import Injectable
-
-__all__ = [
-    "ContainerInjection",
-    "ProviderInjection",
-    "Injectable",
-]
