--- src/dependency/core2/resolution/__init__.py
@@ -1,7 +1,0 @@
-from dependency.core2.resolution.container import Container
-from dependency.core2.resolution.resolver import InjectionResolver
-
-__all__ = [
-    "Container",
-    "InjectionResolver",
-]
