+++ src/example/app/main/__init__.py
@@ -1,5 +1,5 @@
-import time
 import logging
+import time
 from dependency.core import Entrypoint, Container
 
 logger = logging.getLogger("root")
@@ -11,7 +11,7 @@ class MainApplication(Entrypoint):
         # Import all the instances that will be used on the application
         # You can also import the plugins list from the imports file
         from example.app.main.imports import PLUGINS
-        
+
         container = Container.from_dict(config={"config": True}, required=True)
         super().__init__(container, PLUGINS)
