+++ src/example/app/main/imports.py
@@ -1,8 +1,9 @@
+from dependency.core import Plugin
 from example.plugin.base.imports import BasePlugin
 from example.plugin.hardware.imports import HardwarePlugin
 from example.plugin.reporter.imports import ReporterPlugin
 
-PLUGINS = [
+PLUGINS: list[Plugin] = [
     BasePlugin,
     HardwarePlugin,
     ReporterPlugin
