+++ src/example/plugin/base/__init__.py
@@ -4,3 +4,4 @@ from example.plugin.base.settings import BasePluginConfig
 @module()
 class BasePlugin(Plugin):
     meta = PluginMeta(name="BasePlugin", version="0.1.0")
+    config: BasePluginConfig
